Modularise & fix type hint
