+++ backend/streams_explorer/core/extractor/extractor.py
@@ -2,7 +2,7 @@ from typing import List, Optional
 
 from kubernetes.client import V1beta1CronJob
 
-from streams_explorer.core.k8s_app import K8sConfig
+from streams_explorer.models.k8s_config import K8sConfig
 from streams_explorer.models.kafka_connector import KafkaConnector
 from streams_explorer.models.sink import Sink
 from streams_explorer.models.source import Source
